+++ test/git_hub_actions/workflow_test.exs
@@ -54,39 +54,39 @@ defmodule GitHubActions.WorkflowTest do
                           "1.14.5",
                           "1.15.8",
                           "1.16.3",
-                          "1.17.2"
+                          "1.17.3"
                         ],
-                        otp: ["21.3", "22.3", "23.3", "24.3", "25.3", "26.2", "27.0"],
+                        otp: ["21.3", "22.3", "23.3", "24.3", "25.3", "26.2", "27.1"],
                         exclude: [
                           [elixir: "1.10.4", otp: "24.3"],
                           [elixir: "1.10.4", otp: "25.3"],
                           [elixir: "1.10.4", otp: "26.2"],
-                          [elixir: "1.10.4", otp: "27.0"],
+                          [elixir: "1.10.4", otp: "27.1"],
                           [elixir: "1.11.4", otp: "25.3"],
                           [elixir: "1.11.4", otp: "26.2"],
-                          [elixir: "1.11.4", otp: "27.0"],
+                          [elixir: "1.11.4", otp: "27.1"],
                           [elixir: "1.12.3", otp: "21.3"],
                           [elixir: "1.12.3", otp: "25.3"],
                           [elixir: "1.12.3", otp: "26.2"],
-                          [elixir: "1.12.3", otp: "27.0"],
+                          [elixir: "1.12.3", otp: "27.1"],
                           [elixir: "1.13.4", otp: "21.3"],
                           [elixir: "1.13.4", otp: "26.2"],
-                          [elixir: "1.13.4", otp: "27.0"],
+                          [elixir: "1.13.4", otp: "27.1"],
                           [elixir: "1.14.5", otp: "21.3"],
                           [elixir: "1.14.5", otp: "22.3"],
-                          [elixir: "1.14.5", otp: "27.0"],
+                          [elixir: "1.14.5", otp: "27.1"],
                           [elixir: "1.15.8", otp: "21.3"],
                           [elixir: "1.15.8", otp: "22.3"],
                           [elixir: "1.15.8", otp: "23.3"],
-                          [elixir: "1.15.8", otp: "27.0"],
+                          [elixir: "1.15.8", otp: "27.1"],
                           [elixir: "1.16.3", otp: "21.3"],
                           [elixir: "1.16.3", otp: "22.3"],
                           [elixir: "1.16.3", otp: "23.3"],
-                          [elixir: "1.16.3", otp: "27.0"],
-                          [elixir: "1.17.2", otp: "21.3"],
-                          [elixir: "1.17.2", otp: "22.3"],
-                          [elixir: "1.17.2", otp: "23.3"],
-                          [elixir: "1.17.2", otp: "24.3"]
+                          [elixir: "1.16.3", otp: "27.1"],
+                          [elixir: "1.17.3", otp: "21.3"],
+                          [elixir: "1.17.3", otp: "22.3"],
+                          [elixir: "1.17.3", otp: "23.3"],
+                          [elixir: "1.17.3", otp: "24.3"]
                         ]
                       ]
                     ]
@@ -130,22 +130,22 @@ defmodule GitHubActions.WorkflowTest do
                     "runs-on": "ubuntu-20.04",
                     strategy: [
                       matrix: [
-                        elixir: ["1.13.4", "1.14.5", "1.15.8", "1.16.3", "1.17.2"],
-                        otp: ["22.3", "23.3", "24.3", "25.3", "26.2", "27.0"],
+                        elixir: ["1.13.4", "1.14.5", "1.15.8", "1.16.3", "1.17.3"],
+                        otp: ["22.3", "23.3", "24.3", "25.3", "26.2", "27.1"],
                         exclude: [
                           [elixir: "1.13.4", otp: "26.2"],
-                          [elixir: "1.13.4", otp: "27.0"],
+                          [elixir: "1.13.4", otp: "27.1"],
                           [elixir: "1.14.5", otp: "22.3"],
-                          [elixir: "1.14.5", otp: "27.0"],
+                          [elixir: "1.14.5", otp: "27.1"],
                           [elixir: "1.15.8", otp: "22.3"],
                           [elixir: "1.15.8", otp: "23.3"],
-                          [elixir: "1.15.8", otp: "27.0"],
+                          [elixir: "1.15.8", otp: "27.1"],
                           [elixir: "1.16.3", otp: "22.3"],
                           [elixir: "1.16.3", otp: "23.3"],
-                          [elixir: "1.16.3", otp: "27.0"],
-                          [elixir: "1.17.2", otp: "22.3"],
-                          [elixir: "1.17.2", otp: "23.3"],
-                          [elixir: "1.17.2", otp: "24.3"]
+                          [elixir: "1.16.3", otp: "27.1"],
+                          [elixir: "1.17.3", otp: "22.3"],
+                          [elixir: "1.17.3", otp: "23.3"],
+                          [elixir: "1.17.3", otp: "24.3"]
                         ]
                       ]
                     ],
@@ -186,7 +186,7 @@ defmodule GitHubActions.WorkflowTest do
                             "test/support/plts-${{ runner.os }}-${{ matrix.elixir }}-${{ matrix.otp }}-${{ hashFiles(format('{0}{1}', github.workspace, '/mix.lock')) }}"
                         ],
                         if:
-                          "${{ contains(matrix.elixir, '1.17.2') && contains(matrix.otp, '27.0') }}"
+                          "${{ contains(matrix.elixir, '1.17.3') && contains(matrix.otp, '27.1') }}"
                       ],
                       [name: "Get dependencies", run: "mix deps.get"],
                       [name: "Compile dependencies", run: "MIX_ENV=test mix deps.compile"],
@@ -197,38 +197,38 @@ defmodule GitHubActions.WorkflowTest do
                       [
                         name: "Check unused dependencies",
                         if:
-                          "${{ contains(matrix.elixir, '1.17.2') && contains(matrix.otp, '27.0') }}",
+                          "${{ contains(matrix.elixir, '1.17.3') && contains(matrix.otp, '27.1') }}",
                         run: "mix deps.unlock --check-unused"
                       ],
                       [
                         name: "Check code format",
                         if:
-                          "${{ contains(matrix.elixir, '1.17.2') && contains(matrix.otp, '27.0') }}",
+                          "${{ contains(matrix.elixir, '1.17.3') && contains(matrix.otp, '27.1') }}",
                         run: "mix format --check-formatted"
                       ],
                       [
                         name: "Lint code",
                         if:
-                          "${{ contains(matrix.elixir, '1.17.2') && contains(matrix.otp, '27.0') }}",
+                          "${{ contains(matrix.elixir, '1.17.3') && contains(matrix.otp, '27.1') }}",
                         run: "mix credo --strict"
                       ],
                       [
                         name: "Run tests",
                         run: "mix test",
                         if:
-                          "${{ !(contains(matrix.elixir, '1.17.2') && contains(matrix.otp, '27.0')) }}"
+                          "${{ !(contains(matrix.elixir, '1.17.3') && contains(matrix.otp, '27.1')) }}"
                       ],
                       [
                         name: "Run tests with coverage",
                         run: "mix coveralls.github",
                         if:
-                          "${{ contains(matrix.elixir, '1.17.2') && contains(matrix.otp, '27.0') }}"
+                          "${{ contains(matrix.elixir, '1.17.3') && contains(matrix.otp, '27.1') }}"
                       ],
                       [
                         name: "Static code analysis",
                         run: "mix dialyzer --format github --force-check",
                         if:
-                          "${{ contains(matrix.elixir, '1.17.2') && contains(matrix.otp, '27.0') }}"
+                          "${{ contains(matrix.elixir, '1.17.3') && contains(matrix.otp, '27.1') }}"
                       ]
                     ]
                   ]